python/requirements.txt
feedparser
Mastodon.py
pytz